Fix OpenShiftOidcWebSocketIT execution on IBM PE/Z
(cherry picked from commit 8cc42c8e41bdcafe4226be2af60ce2fab7897a05)
