+++ websockets/websocket-next-oidc/src/test/java/io/quarkus/ts/websockets/next/OpenShiftOidcWebSocketIT.java
@@ -1,7 +1,17 @@
 package io.quarkus.ts.websockets.next;
 
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM;
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM_BASE_PATH;
+import static io.quarkus.test.bootstrap.KeycloakService.DEFAULT_REALM_FILE;
+
+import io.quarkus.test.bootstrap.KeycloakService;
 import io.quarkus.test.scenarios.OpenShiftScenario;
+import io.quarkus.test.services.KeycloakContainer;
 
 @OpenShiftScenario
-public class OpenShiftOidcWebSocketIT extends OidcWebSocketIT {
+public class OpenShiftOidcWebSocketIT extends AbstractOidcWebSocketIT {
+
+    @KeycloakContainer(command = { "start-dev", "--import-realm", "--features=token-exchange" }, image = "${rhbk.image}")
+    static KeycloakService keycloak = new KeycloakService(DEFAULT_REALM_FILE, DEFAULT_REALM, DEFAULT_REALM_BASE_PATH);
+
 }
